pyproject.toml
[project]
name = "ongbd-wikipedia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "neo4j==1.7.6",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "wikipedia-api>=0.8.1",
]
